Version 1.1
Separate sounds when keys are pressed and released, and for the shift key.

Better (manual) carriage return sound.

Limit the speed at which you can type.

If you keep typing past the bell and hit the end of the line (80 characters), the carriage stops instead of continuing. In other words there is now a hard line length limit.

More punctuation characters are handled (these are based on my UK Mac keyboard; unless you have the same keyboard some characters will probably be "wrong" for you.).

Random variation of colour to simulate variations in key pressure and uneven ribbon ink.

Hide the informational text when you start typing, partly to give more space, but mostly to dodge the problem of trying to work out where to position the typewriter output.