version 1.15
* A better algorithm for mapping Brokenness to vertical offset.
Rather than having a fixed offset for each character, and simply adding more
random offset when Brokenness is increased, we now choose an offset and threshold
for each character. If Brokenness is above the threshold for that char, it will be
offset by its consistent amount. This offset may then be randomly boosted (or not)
by 1 or 2px at high Brokenness levels.

* Brokenness now maxes at 99%, given that 100% brokenness would imply being
completely non-functional.

* Show the values of the sliders as numbers below them.

* Reduce max inkiness.